// src/infrastructure/enums/http-name.enum.ts
export enum HTTP_NAME {
  BAD_REQUEST = "BAD_REQUEST",
  UNAUTHORIZED = "UNAUTHORIZED",
  PAYMENT_REQUIRED = "PAYMENT_REQUIRED",
  FORBIDDEN = "FORBIDDEN",
  NOT_FOUND = "NOT_FOUND",
  METHOD_NOT_ALLOWED = "METHOD_NOT_ALLOWED",
  NOT_ACCEPTABLE = "NOT_ACCEPTABLE",
  PROXY_AUTHENTICATION_REQUIRED = "PROXY_AUTHENTICATION_REQUIRED",
  REQUEST_TIMEOUT = "REQUEST_TIMEOUT",
  CONFLICT = "CONFLICT",
  GONE = "GONE",
  LENGTH_REQUIRED = "LENGTH_REQUIRED",
  PRECONDITION_FAILED = "PRECONDITION_FAILED",
  PAYLOAD_TOO_LARGE = "PAYLOAD_TOO_LARGE",
  REQUEST_URI_TOO_LONG = "REQUEST_URI_TOO_LONG",
  UNSUPPORTED_MEDIA_TYPE = "UNSUPPORTED_MEDIA_TYPE",
  REQUESTED_RANGE_NOT_SATISFIABLE = "REQUESTED_RANGE_NOT_SATISFIABLE",
  EXPECTATION_FAILED = "EXPECTATION_FAILED",
  IM_A_TEAPOT = "IM_A_TEAPOT",
  MISDIRECTED_REQUEST = "MISDIRECTED_REQUEST",
  UNPROCESSABLE_ENTITY = "UNPROCESSABLE_ENTITY",
  LOCKED = "LOCKED",
  FAILED_DEPENDENCY = "FAILED_DEPENDENCY",
  UPGRADE_REQUIRED = "UPGRADE_REQUIRED",
  PRECONDITION_REQUIRED = "PRECONDITION_REQUIRED",
  TOO_MANY_REQUESTS = "TOO_MANY_REQUESTS",
  REQUEST_HEADER_FIELDS_TOO_LARGE = "REQUEST_HEADER_FIELDS_TOO_LARGE",
  CONNECTION_CLOSED_WITHOUT_RESPONSE = "CONNECTION_CLOSED_WITHOUT_RESPONSE",
  UNAVAILABLE_FOR_LEGAL_REASONS = "UNAVAILABLE_FOR_LEGAL_REASONS",
  CLIENT_CLOSED_REQUEST = "CLIENT_CLOSED_REQUEST",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  NOT_IMPLEMENTED = "NOT_IMPLEMENTED",
  BAD_GATEWAY = "BAD_GATEWAY",
  SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE",
  GATEWAY_TIMEOUT = "GATEWAY_TIMEOUT",
  HTTP_VERSION_NOT_SUPPORTED = "HTTP_VERSION_NOT_SUPPORTED",
  VARIANT_ALSO_NEGOTIATES = "VARIANT_ALSO_NEGOTIATES",
  INSUFFICIENT_STORAGE = "INSUFFICIENT_STORAGE",
  LOOP_DETECTED = "LOOP_DETECTED",
  NOT_EXTENDED = "NOT_EXTENDED",
  NETWORK_AUTHENTICATION_REQUIRED = "NETWORK_AUTHENTICATION_REQUIRED",
  NETWORK_CONNECTION_TIMEOUT_ERROR = "NETWORK_CONNECTION_TIMEOUT_ERROR",
}